+++ ryft-ui/src/main/java/com/ryftpay/android/ui/service/GooglePayService.kt
@@ -2,7 +2,7 @@ package com.ryftpay.android.ui.service
 
 import android.app.Activity
 import com.google.android.gms.tasks.Task
-import com.ryftpay.android.ui.service.request.LoadPaymentDataRequest
+import com.ryftpay.android.ui.service.request.googlepay.LoadPaymentDataRequest
 
 internal interface GooglePayService {
     fun isReadyToPay(): Task<Boolean>
