+++ ryft-ui/src/main/java/com/ryftpay/android/ui/service/request/googlepay/LoadPaymentDataRequest.kt
@@ -1,4 +1,4 @@
-package com.ryftpay.android.ui.service.request
+package com.ryftpay.android.ui.service.request.googlepay
 
 internal data class LoadPaymentDataRequest(
     val merchantInfo: MerchantInfo,
