+++ ryft-ui/src/main/java/com/ryftpay/android/ui/service/request/googlepay/RyftTokenizationSpecification.kt
@@ -1,4 +1,4 @@
-package com.ryftpay.android.ui.service.request
+package com.ryftpay.android.ui.service.request.googlepay
 
 import com.ryftpay.android.core.model.api.RyftPublicApiKey
