+++ ryft-ui/src/test/java/com/ryftpay/android/ui/service/request/googlepay/MerchantInfoTest.kt
@@ -1,4 +1,4 @@
-package com.ryftpay.android.ui.service.request
+package com.ryftpay.android.ui.service.request.googlepay
 
 import com.ryftpay.android.ui.TestData.MERCHANT_NAME
 import org.amshove.kluent.shouldBeEqualTo
